NOT using sql for unread and unacknowledged comment counts
git-svn-id: svn://pele.cx/site/story/trunk@402 7189bf43-9712-0410-8374-e18a92618762

+++ app/controllers/chapters_controller.rb
@@ -87,9 +87,11 @@ class ChaptersController < ApplicationController
   def update
     @chapter = Chapter.find(params[:id])
 
-    unless params[:file].size <= 0
-      Paragraph.delete_all ["chapter_id = ?", @chapter.id]
-      process_file(params[:file],@chapter.id)
+    if params[:file]
+      unless params[:file].size <= 0
+        Paragraph.delete_all ["chapter_id = ?", @chapter.id]
+        process_file(params[:file],@chapter.id)
+      end
     end
 
     oldstatus = @chapter.status
